Extract common integration helper code
